+++ doc/source/conf.py
@@ -36,8 +36,9 @@ import tbmodels
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc', 'sphinx.ext.mathjax', 'sphinx.ext.intersphinx', 'sphinx.ext.viewcode', 'sphinx_click.ext',
-    'sphinx.ext.napoleon', 'IPython.sphinxext.ipython_console_highlighting', 'IPython.sphinxext.ipython_directive'
+    'sphinx.ext.autodoc', 'sphinx.ext.mathjax', 'sphinx.ext.intersphinx', 'sphinx.ext.viewcode',
+    'sphinx_click.ext', 'sphinx.ext.napoleon', 'IPython.sphinxext.ipython_console_highlighting',
+    'IPython.sphinxext.ipython_directive'
 ]
 
 intersphinx_mapping = {