+++ examples/cmdline/symmetrization/generate_symmetries.py
@@ -60,7 +60,9 @@ def spin_reps(prep):
             else:
                 n /= np.linalg.norm(n)
                 # rotation followed by reflection:
-                spin = np.round(np.dot(D12(n[0], n[1], n[2], np.pi), D12(n[0], n[1], n[2], theta)), 15)
+                spin = np.round(
+                    np.dot(D12(n[0], n[1], n[2], np.pi), D12(n[0], n[1], n[2], theta)), 15
+                )
         else:  # case of inversion (does not do anything to spin)
             spin = D12(0, 0, 0, 0)
     return np.array(spin)
@@ -73,11 +75,15 @@ if __name__ == '__main__':
 
     # set up symmetry operations
     time_reversal = sr.SymmetryOperation(
-        rotation_matrix=np.eye(3), repr_matrix=np.kron([[0, -1j], [1j, 0]], np.eye(7)), repr_has_cc=True
+        rotation_matrix=np.eye(3),
+        repr_matrix=np.kron([[0, -1j], [1j, 0]], np.eye(7)),
+        repr_has_cc=True
     )
 
     structure = mg.Structure(
-        lattice=model_nosym.uc, species=['In', 'As'], coords=np.array([[0, 0, 0], [0.25, 0.25, 0.25]])
+        lattice=model_nosym.uc,
+        species=['In', 'As'],
+        coords=np.array([[0, 0, 0], [0.25, 0.25, 0.25]])
     )
 
     # get real-space representations