+++ tbmodels/_tb_model.py
@@ -28,6 +28,10 @@ from . import _check_compatibility
 from . import _sparse_matrix as sp
 from .kdotp import KdotpModel
 
+__all__ = ('Model', )
+
+HoppingType = ty.Dict[ty.Tuple[int, ...], ty.Any]
+
 
 @export
 @subscribe_hdf5('tbmodels.model', check_on_load=False)
@@ -71,12 +75,12 @@ class Model(HDF5Enabled):
     def __init__(
         self,
         *,
-        on_site: ty.Optional[ty.Collection[float]] = None,
-        hop: ty.Optional[ty.Mapping[ty.Tuple[int, ...], ty.Any]] = None,
+        on_site: ty.Optional[ty.Sequence[float]] = None,
+        hop: ty.Optional[HoppingType] = None,
         size: ty.Optional[int] = None,
         dim: ty.Optional[int] = None,
         occ: ty.Optional[int] = None,
-        pos: ty.Optional[ty.Collection[ty.Collection[float]]] = None,
+        pos: ty.Optional[ty.Sequence[ty.Sequence[float]]] = None,
         uc: ty.Optional[np.ndarray] = None,
         contains_cc: bool = True,
         sparse: bool = False
@@ -151,7 +155,10 @@ class Model(HDF5Enabled):
         """
         # The double-constructor is needed to avoid a double-constructor in the sparse to-array
         # but still allow for the dtype argument.
-        hop = {tuple(key): self._matrix_type(self._matrix_type(value), dtype=complex) for key, value in hop.items()}
+        hop = {
+            tuple(key): self._matrix_type(self._matrix_type(value), dtype=complex)
+            for key, value in hop.items()
+        }
 
         # positions
         if pos is None:
@@ -182,9 +189,8 @@ class Model(HDF5Enabled):
         if on_site is not None:
             if len(on_site) != self.size:
                 raise ValueError(
-                    'The number of on-site energies {0} does not match the size of the system {1}'.format(
-                        len(on_site), self.size
-                    )
+                    'The number of on-site energies {0} does not match the size of the system {1}'.
+                    format(len(on_site), self.size)
                 )
             self.hop[self._zero_vec] += 0.5 * self._matrix_type(np.diag(on_site))
 
@@ -217,11 +223,21 @@ class Model(HDF5Enabled):
         Reduce the full hoppings representation (with cc) to the reduced one (without cc, zero-terms halved).
         """
         # Consistency checks
+        failed_R = []
         for R, mat in hop.items():
-            if la.norm(mat - hop.get(tuple(-x for x in R), np.zeros(mat.shape)).T.conjugate()) > 1e-12:
-                raise ValueError(
-                    'The provided hoppings do not correspond to a hermitian Hamiltonian. hoppings[-R] = hoppings[R].H is not fulfilled.'
+            diff_norm = la.norm(
+                mat - hop.get(tuple(-x for x in R), np.zeros(mat.shape)).T.conjugate()
+            )
+            if diff_norm > 1e-12:
+                failed_R.append((R, diff_norm))
+        if failed_R:
+            raise ValueError(
+                'The provided hoppings do not correspond to a hermitian Hamiltonian. hoppings[-R] = hoppings[R].H is not fulfilled for the following values:\n'
+                + '\n'.join(
+                    f'R={R}, delta_norm={diff_norm}'
+                    for R, diff_norm in sorted(failed_R, key=lambda val: -val[1])
                 )
+            )
 
         res = dict()
         for R, mat in hop.items():
@@ -235,11 +251,11 @@ class Model(HDF5Enabled):
                 res[R] = 0.5 * mat
         return res
 
-    def _map_hop_positive_R(self, hop):
+    def _map_hop_positive_R(self, hop: HoppingType) -> HoppingType:
         """
         Maps hoppings with a negative first non-zero index in R to their positive counterpart.
         """
-        new_hop = co.defaultdict(self._empty_matrix)
+        new_hop: HoppingType = co.defaultdict(self._empty_matrix)
         for R, mat in hop.items():
             try:
                 if R[np.nonzero(R)[0][0]] > 0:
@@ -263,7 +279,9 @@ class Model(HDF5Enabled):
         for h_mat in self.hop.values():
             if not h_mat.shape == (self.size, self.size):
                 raise ValueError(
-                    'Hopping matrix of shape {0} found, should be ({1},{1}).'.format(h_mat.shape, self.size)
+                    'Hopping matrix of shape {0} found, should be ({1},{1}).'.format(
+                        h_mat.shape, self.size
+                    )
                 )
 
     def _check_dim(self):
@@ -271,7 +289,8 @@ class Model(HDF5Enabled):
         for key in self.hop.keys():
             if len(key) != self.dim:
                 raise ValueError(
-                    'The length of R = {0} does not match the dimensionality of the system ({1})'.format(key, self.dim)
+                    'The length of R = {0} does not match the dimensionality of the system ({1})'.
+                    format(key, self.dim)
                 )
         if self.uc is not None:
             if self.uc.shape != (self.dim, self.dim):
@@ -285,7 +304,7 @@ class Model(HDF5Enabled):
     def from_hop_list(
         cls,
         *,
-        hop_list: ty.Iterable[ty.Collection[ty.Union[complex, int, ty.Collection[int]]]] = (),
+        hop_list: ty.Iterable[ty.Tuple[complex, int, int, ty.Tuple[int, ...]]] = (),
         size: ty.Optional[int] = None,
         **kwargs
     ) -> "Model":
@@ -311,7 +330,9 @@ class Model(HDF5Enabled):
             try:
                 size = len(kwargs['on_site'])
             except KeyError:
-                raise ValueError('No on-site energies and no size given. The size of the system cannot be determined.')
+                raise ValueError(
+                    'No on-site energies and no size given. The size of the system cannot be determined.'
+                )
 
         class _hop:
             """
@@ -329,15 +350,17 @@ class Model(HDF5Enabled):
 
         # create data, row_idx, col_idx for setting up the CSR matrices
         hop_list_dict: ty.Mapping[ty.Tuple[int, ...], _hop] = co.defaultdict(_hop)
-        R: ty.Collection[int]
-        for t, i, j, R in hop_list:  # type: ignore
+        R: ty.Tuple[int, ...]
+        for t, i, j, R in hop_list:
             R_vec = tuple(R)
             hop_list_dict[R_vec].append(t, i, j)
 
         # creating CSR matrices
         hop_dict = dict()
         for key, val in hop_list_dict.items():
-            hop_dict[key] = sp.csr((val.data, (val.row_idx, val.col_idx)), dtype=complex, shape=(size, size))
+            hop_dict[key] = sp.csr((val.data, (val.row_idx, val.col_idx)),
+                                   dtype=complex,
+                                   shape=(size, size))
 
         return cls(size=size, hop=hop_dict, **kwargs)
 
@@ -367,10 +390,11 @@ class Model(HDF5Enabled):
             orbital_b = int(entry[4]) - 1
             # test consistency of orbital numbers
             if not ignore_orbital_order:
-                if not (orbital_a == i % num_wann) and (orbital_b == (i % num_wann_square) // num_wann):
+                if not (orbital_a == i %
+                        num_wann) and (orbital_b == (i % num_wann_square) // num_wann):
                     raise ValueError("Inconsistent orbital numbers in line '{}'".format(line))
-            return [(float(entry[5]) + 1j * float(entry[6])) / (deg_pts[i // num_wann_square]), orbital_a, orbital_b,
-                    [int(x) for x in entry[:3]]]
+            return [(float(entry[5]) + 1j * float(entry[6])) / (deg_pts[i // num_wann_square]),
+                    orbital_a, orbital_b, [int(x) for x in entry[:3]]]
 
         # skip random empty lines
         lines_nonempty = (l for l in iterator if l.strip())
@@ -434,7 +458,8 @@ class Model(HDF5Enabled):
         if self._zero_vec in self.hop.keys():
             lines.extend(
                 self._mat_to_hr(
-                    self._zero_vec, self.hop[self._zero_vec] + self.hop[self._zero_vec].conjugate().transpose()
+                    self._zero_vec,
+                    self.hop[self._zero_vec] + self.hop[self._zero_vec].conjugate().transpose()
                 )
             )
         # positive
@@ -454,9 +479,8 @@ class Model(HDF5Enabled):
         for j, column in enumerate(mat):
             for i, t in enumerate(column):
                 lines.append(
-                    '{0[0]:>5}{0[1]:>5}{0[2]:>5}{1:>5}{2:>5}{3.real:>22.14f}{3.imag:>22.14f}'.format(
-                        R, i + 1, j + 1, t
-                    )
+                    '{0[0]:>5}{0[1]:>5}{0[2]:>5}{1:>5}{2:>5}{3.real:>22.14f}{3.imag:>22.14f}'.
+                    format(R, i + 1, j + 1, t)
                 )
         return lines
 
@@ -567,14 +591,16 @@ class Model(HDF5Enabled):
                         p_reduced = la.solve(kwargs['uc'].T, np.array(p).T).T
                         T_base = np.floor(p_reduced)
                         all_atom_pos = np.array([
-                            kwargs['uc'].T @ (T_base + T_shift) + atom_pos for atom_pos in atom_pos_cartesian
+                            kwargs['uc'].T @ (T_base + T_shift) + atom_pos
+                            for atom_pos in atom_pos_cartesian
                             for T_shift in itertools.product([-1, 0, 1], repeat=3)
                         ])
                         distances = la.norm(p - all_atom_pos, axis=-1)
                         pos_cartesian.append(all_atom_pos[np.argmin(distances)])
                 else:
                     raise ValueError(
-                        "Invalid value '{}' for 'pos_kind', must be 'wannier' or 'nearest_atom'".format(pos_kind)
+                        "Invalid value '{}' for 'pos_kind', must be 'wannier' or 'nearest_atom'".
+                        format(pos_kind)
                     )
                 kwargs['pos'] = la.solve(kwargs['uc'].T, np.array(pos_cartesian).T).T
 
@@ -596,12 +622,16 @@ class Model(HDF5Enabled):
                             # Explicitly catching this because 'remap_hoppings'
                             # is also a generator.
                             except StopIteration as exc:
-                                raise ValueError("The 'wsvec_generator' stopped prematurely.") from exc
+                                raise ValueError(
+                                    "The 'wsvec_generator' stopped prematurely."
+                                ) from exc
                             T_list = wsvec_generator.send((orbital_1, orbital_2, tuple(R)))
                             N = len(T_list)
                             for T in T_list:
                                 # not using numpy here increases performance
-                                yield (t / N, orbital_1, orbital_2, tuple(r + t for r, t in zip(R, T)))
+                                yield (
+                                    t / N, orbital_1, orbital_2, tuple(r + t for r, t in zip(R, T))
+                                )
 
                     hop_entries = remap_hoppings(hop_entries)
                     return cls.from_hop_list(size=num_wann, hop_list=hop_entries, **kwargs)
@@ -779,8 +809,9 @@ class Model(HDF5Enabled):
                 if i == j:
                     # handled above
                     continue
-                kwant_sys[kwant.builder.HoppingKind(self._zero_vec, kwant_sublattices[i],
-                                                    kwant_sublattices[j])] = on_site_mat[np.ix_(s1.indices, s2.indices)]
+                kwant_sys[kwant.builder.HoppingKind(
+                    self._zero_vec, kwant_sublattices[i], kwant_sublattices[j]
+                )] = on_site_mat[np.ix_(s1.indices, s2.indices)]
 
         # R != 0 terms
         for R, mat in self.hop.items():
@@ -793,10 +824,12 @@ class Model(HDF5Enabled):
                 for j, s2 in enumerate(sublattices):
                     sub_matrix = mat[np.ix_(s1.indices, s2.indices)]
                     # TODO: check "signs"
-                    kwant_sys[kwant.builder.HoppingKind(minus_R, kwant_sublattices[i],
-                                                        kwant_sublattices[j])] = sub_matrix
-                    kwant_sys[kwant.builder.HoppingKind(R, kwant_sublattices[j],
-                                                        kwant_sublattices[i])] = np.transpose(np.conj(sub_matrix))
+                    kwant_sys[kwant.builder.HoppingKind(
+                        minus_R, kwant_sublattices[i], kwant_sublattices[j]
+                    )] = sub_matrix
+                    kwant_sys[
+                        kwant.builder.HoppingKind(R, kwant_sublattices[j], kwant_sublattices[i])
+                    ] = np.transpose(np.conj(sub_matrix))
         return kwant_sys
 
     def _get_sublattices(self):
@@ -817,7 +850,7 @@ class Model(HDF5Enabled):
                 sublattices.append(Sublattice(pos=p_orb, indices=[i]))
         return sublattices
 
-    def construct_kdotp(self, k: ty.Collection[float], order: int):
+    def construct_kdotp(self, k: ty.Sequence[float], order: int):
         """
         Construct a k.p model around a given k-point. This is done by explicitly
         evaluating the derivatives which make up the Taylor expansion of the k.p
@@ -835,8 +868,6 @@ class Model(HDF5Enabled):
             The order (sum of powers) to which the Taylor expansion is
             performed.
         """
-        # foo : ty.Collection[int] = (1, 2, 3)
-        # taylor_coefficients : ty.Dict[ty.Tuple[int, ...], ty.Any] = dict()
         taylor_coefficients = dict()
         if order < 0:
             raise ValueError('The order for the k.p model must be positive.')
@@ -848,9 +879,10 @@ class Model(HDF5Enabled):
             taylor_coefficients[k_powers] = (
                 (2j * np.pi)**curr_order / np.prod(factorial(k_powers, exact=True))
             ) * sum((
-                np.prod(np.array(R)**np.array(k_powers)) * np.exp(2j * np.pi * np.dot(k, R)) * self._array_cast(mat) +
-                np.prod((-np.array(R))**np.array(k_powers)) * np.exp(-2j * np.pi * np.dot(k, R)) *
-                self._array_cast(mat).T.conj() for R, mat in self.hop.items()
+                np.prod(np.array(R)**np.array(k_powers)) * np.exp(2j * np.pi * np.dot(k, R)) *
+                self._array_cast(mat) + np.prod((-np.array(R))**np.array(k_powers)) *
+                np.exp(-2j * np.pi * np.dot(k, R)) * self._array_cast(mat).T.conj()
+                for R, mat in self.hop.items()
             ), np.zeros((self.size, self.size), dtype=complex))
         return KdotpModel(taylor_coefficients=taylor_coefficients)
 
@@ -890,8 +922,9 @@ class Model(HDF5Enabled):
             for group in tb_model_group['hop'].values():
                 R = tuple(group['R'])
                 if new_kwargs['sparse']:
-                    new_kwargs['hop'][R] = sp.csr((group['data'], group['indices'], group['indptr']),
-                                                  shape=group['shape'])
+                    new_kwargs['hop'][R] = sp.csr(
+                        (group['data'], group['indices'], group['indptr']), shape=group['shape']
+                    )
                 else:
                     new_kwargs['hop'][R] = np.array(group['mat'])
             new_kwargs['contains_cc'] = False
@@ -920,9 +953,8 @@ class Model(HDF5Enabled):
 
     def __repr__(self):
         return ' '.join(
-            'tbmodels.Model(hop={1}, pos={0.pos!r}, uc={0.uc!r}, occ={0.occ}, contains_cc=False)'.format(
-                self, dict(self.hop)
-            ).replace('\n', ' ').replace('array', 'np.array').split()
+            'tbmodels.Model(hop={1}, pos={0.pos!r}, uc={0.uc!r}, occ={0.occ}, contains_cc=False)'.
+            format(self, dict(self.hop)).replace('\n', ' ').replace('array', 'np.array').split()
         )
 
     #---------------- BASIC FUNCTIONALITY ----------------------------------#
@@ -931,7 +963,7 @@ class Model(HDF5Enabled):
         """An array containing the reciprocal lattice vectors as rows."""
         return None if self.uc is None else 2 * np.pi * la.inv(self.uc).T
 
-    def hamilton(self, k: ty.Collection[float], convention: int = 2) -> np.ndarray:
+    def hamilton(self, k: ty.Sequence[float], convention: int = 2) -> np.ndarray:
         """
         Calculates the Hamilton matrix for a given k-point.
 
@@ -946,17 +978,20 @@ class Model(HDF5Enabled):
             Valid choices are 1 or 2.
         """
         if convention not in [1, 2]:
-            raise ValueError("Invalid value '{}' for 'convention': must be either '1' or '2'".format(convention))
+            raise ValueError(
+                "Invalid value '{}' for 'convention': must be either '1' or '2'".format(convention)
+            )
         k = np.array(k, ndmin=1)
-        H = sum((self._array_cast(hop) * np.exp(2j * np.pi * np.dot(R, k)) for R, hop in self.hop.items()),
-                np.zeros((self.size, self.size), dtype=complex))
+        H = sum((
+            self._array_cast(hop) * np.exp(2j * np.pi * np.dot(R, k)) for R, hop in self.hop.items()
+        ), np.zeros((self.size, self.size), dtype=complex))
         H += H.conjugate().T
         if convention == 1:
             pos_exponential = np.array([[np.exp(2j * np.pi * np.dot(p, k)) for p in self.pos]])
             H = pos_exponential.conjugate().transpose() * H * pos_exponential
         return H
 
-    def eigenval(self, k: ty.Collection[float]) -> np.ndarray:
+    def eigenval(self, k: ty.Sequence[float]) -> np.ndarray:
         """
         Returns the eigenvalues at a given k point.
 
@@ -968,7 +1003,7 @@ class Model(HDF5Enabled):
         return la.eigvalsh(self.hamilton(k))
 
     #-------------------MODIFYING THE MODEL ----------------------------#
-    def add_hop(self, overlap: complex, orbital_1: int, orbital_2: int, R: ty.Collection[int]):
+    def add_hop(self, overlap: complex, orbital_1: int, orbital_2: int, R: ty.Sequence[int]):
         r"""
         Adds a hopping term with a given overlap (hopping strength) from
         ``orbital_2`` (:math:`o_2`), which lies in the unit cell pointed
@@ -1007,7 +1042,11 @@ class Model(HDF5Enabled):
         """
         R = tuple(R)
         if len(R) != self.dim:
-            raise ValueError('Dimension of R ({}) does not match the model dimension ({})'.format(len(R), self.dim))
+            raise ValueError(
+                'Dimension of R ({}) does not match the model dimension ({})'.format(
+                    len(R), self.dim
+                )
+            )
 
         mat = np.zeros((self.size, self.size), dtype=complex)
         nonzero_idx = np.nonzero(R)[0]
@@ -1021,7 +1060,7 @@ class Model(HDF5Enabled):
             mat[orbital_2, orbital_1] += overlap.conjugate()
         self.hop[R] += self._matrix_type(mat)
 
-    def add_on_site(self, on_site: ty.Collection[float]):
+    def add_on_site(self, on_site: ty.Sequence[float]):
         """
         Adds on-site energy to the orbitals. This adds to the existing
         on-site energy, and does not erase it.
@@ -1033,7 +1072,9 @@ class Model(HDF5Enabled):
         """
         if self.size != len(on_site):
             raise ValueError(
-                'The number of on-site energy terms should be {}, but is {}.'.format(self.size, len(on_site))
+                'The number of on-site energy terms should be {}, but is {}.'.format(
+                    self.size, len(on_site)
+                )
             )
         for orbital, energy in enumerate(on_site):
             self.add_hop(energy / 2., orbital, orbital, self._zero_vec)
@@ -1085,9 +1126,18 @@ class Model(HDF5Enabled):
     #---- arithmetic operations ----#
     @property
     def _input_kwargs(self):
-        return dict(hop=self.hop, pos=self.pos, occ=self.occ, uc=self.uc, contains_cc=False, sparse=self._sparse)
+        return dict(
+            hop=self.hop,
+            pos=self.pos,
+            occ=self.occ,
+            uc=self.uc,
+            contains_cc=False,
+            sparse=self._sparse
+        )
 
-    def symmetrize(self, symmetries: ty.Sequence[sr.SymmetryOperation], full_group: bool = False) -> "Model":
+    def symmetrize(
+        self, symmetries: ty.Sequence[sr.SymmetryOperation], full_group: bool = False
+    ) -> "Model":
         """
         Returns a model which is symmetrized w.r.t. the given
         symmetries. This is done by performing a group average over the
@@ -1104,7 +1154,8 @@ class Model(HDF5Enabled):
         """
         if full_group:
             new_model = self._apply_operation(symmetries[0])
-            return 1 / len(symmetries) * sum((self._apply_operation(s) for s in symmetries[1:]), new_model)
+            return 1 / len(symmetries) * sum((self._apply_operation(s)
+                                              for s in symmetries[1:]), new_model)
         else:
             new_model = self
             for sym in symmetries:
@@ -1124,7 +1175,9 @@ class Model(HDF5Enabled):
         # apply symmetry operation on sublattice positions
         sublattices = self._get_sublattices()
 
-        new_sublattice_pos = [symmetry_operation.real_space_operator.apply(latt.pos) for latt in sublattices]
+        new_sublattice_pos = [
+            symmetry_operation.real_space_operator.apply(latt.pos) for latt in sublattices
+        ]
 
         # match to a known sublattice position to determine the shift vector
         uc_shift = []
@@ -1137,16 +1190,20 @@ class Model(HDF5Enabled):
                 if any(np.isclose(new_pos - shift, latt.pos).all() for latt in sublattices):
                     valid_shifts.append(tuple(shift))
             if not valid_shifts:
-                raise ValueError('New position {} does not match any known sublattice'.format(new_pos))
+                raise ValueError(
+                    'New position {} does not match any known sublattice'.format(new_pos)
+                )
             if len(valid_shifts) > 1:
                 raise ValueError(
-                    'Ambiguity error: New position {} matches more than one known sublattice'.format(new_pos)
+                    'Ambiguity error: New position {} matches more than one known sublattice'.
+                    format(new_pos)
                 )
             uc_shift.append(valid_shifts[0])
 
         # setting up the indices to slice the hopping matrices
-        hop_shifts_idx: ty.Dict[ty.Tuple[int, ...], ty.Tuple[ty.List[int],
-                                                             ty.List[int]]] = co.defaultdict(lambda: ([], []))
+        hop_shifts_idx: ty.Dict[ty.Tuple[int, ...],
+                                ty.Tuple[ty.List[int],
+                                         ty.List[int]]] = co.defaultdict(lambda: ([], []))
         for (i, Ti), (j, Tj) in itertools.product(enumerate(uc_shift), repeat=2):
             shift = tuple(np.array(Tj) - np.array(Ti))
             for idx1, idx2 in itertools.product(sublattices[i].indices, sublattices[j].indices):
@@ -1154,9 +1211,11 @@ class Model(HDF5Enabled):
                 hop_shifts_idx[shift][1].append(idx2)
 
         # create hoppings with shifted R (by uc_shift[j] - uc_shift[i])
-        new_hop: ty.Dict[ty.Tuple[int, ...], ty.Any] = co.defaultdict(self._empty_matrix)
+        new_hop: HoppingType = co.defaultdict(self._empty_matrix)
         for R, mat in self.hop.items():
-            R_transformed = np.array(np.rint(np.dot(symmetry_operation.rotation_matrix, R)), dtype=int)
+            R_transformed = np.array(
+                np.rint(np.dot(symmetry_operation.rotation_matrix, R)), dtype=int
+            )
             for shift, (idx1, idx2) in hop_shifts_idx.items():
                 new_R = tuple(np.array(R_transformed) + np.array(shift))
                 new_hop[new_R][idx1, idx2] += mat[idx1, idx2]
@@ -1181,7 +1240,10 @@ class Model(HDF5Enabled):
             Orbital indices that will be in the resulting model.
         """
         new_pos = self.pos[tuple(slice_idx), :]
-        new_hop = {key: np.array(val)[np.ix_(slice_idx, slice_idx)] for key, val in self.hop.items()}
+        new_hop = {
+            key: np.array(val)[np.ix_(slice_idx, slice_idx)]
+            for key, val in self.hop.items()
+        }
         return Model(**co.ChainMap(dict(hop=new_hop, pos=new_pos), self._input_kwargs))
 
     @classmethod
@@ -1239,6 +1301,275 @@ class Model(HDF5Enabled):
 
         return cls(dim=new_dim, uc=new_uc, pos=new_pos, occ=new_occ, hop=new_hop, contains_cc=False)
 
+    def supercell(self, size: ty.Sequence[int]) -> "Model":  # pylint: disable=too-many-locals
+        """Generate a model for a supercell of the current unit cell.
+
+        Parameters
+        ----------
+        size :
+            The size of the supercell, given as integer multiples of the
+            current lattice vectors
+        """
+        size_array = np.array(size).astype(dtype=int, casting='safe')
+        if size_array.shape != (self.dim, ):
+            raise ValueError(
+                "The given 'size' has incorrect shape {}, should be {}.".format(
+                    size_array.shape, (self.dim, )
+                )
+            )
+        volume_multiplier = np.prod(size_array)
+        new_occ = None if self.occ is None else volume_multiplier * self.occ
+        if self.uc is None:
+            new_uc = None
+        else:
+            new_uc = (self.uc.T * size_array).T
+
+        # the new positions, normalized to the supercell
+        new_pos: ty.List[np.ndarray] = []
+        reduced_pos = np.array([p / size_array for p in self.pos])
+        uc_offsets = list(
+            np.array(offset) for offset in itertools.product(*[range(n) for n in size_array])
+        )
+        for current_uc_offset in uc_offsets:
+            new_pos.extend(reduced_pos + (current_uc_offset / size_array))
+
+        new_size = self.size * volume_multiplier
+        new_hop: HoppingType = co.defaultdict(lambda: np.zeros((new_size, new_size), dtype=complex))
+
+        # Can be used to get the orbital offset of a given unit cell
+        # by taking the inner product with the unit cell position.
+        uc_idx_multiplier = np.array([
+            np.prod(size[i:], dtype=int) for i in range(1,
+                                                        len(size) + 1)
+        ]) * self.size
+
+        for uc1_idx, uc1_pos in enumerate(uc_offsets):
+            uc1_idx_offset = uc1_idx * self.size
+
+            for R, hop_mat in self.hop.items():
+                hop_mat = self._array_cast(hop_mat)
+
+                # position of the uc of orbital 2, not mapped inside supercell
+                full_uc2_pos = uc1_pos + R
+                # mapped into the supercell
+                uc2_pos = full_uc2_pos % size_array
+                uc2_idx_offset = np.inner(uc_idx_multiplier, uc2_pos)
+
+                # R in terms of supercells
+                new_R = np.array(np.floor(full_uc2_pos / size_array), dtype=int)
+
+                new_hop[tuple(new_R)][uc1_idx_offset:uc1_idx_offset + self.size,
+                                      uc2_idx_offset:uc2_idx_offset + self.size] += hop_mat
+
+        return Model(
+            **co.ChainMap(
+                dict(
+                    hop=new_hop,
+                    occ=new_occ,
+                    uc=new_uc,
+                    size=new_size,
+                    pos=new_pos,
+                    contains_cc=False
+                ), self._input_kwargs
+            )
+        )
+
+    def fold_model(  # pylint: disable=too-many-locals # noqa:C001
+        self,
+        *,
+        new_unit_cell: ty.Sequence[ty.Sequence[float]],
+        unit_cell_offset: ty.Sequence[float] = (0, 0, 0),
+        position_tolerance: float = 1e-3,
+        orbital_labels: ty.Sequence[ty.Hashable],
+        target_indices: ty.Optional[ty.Sequence[int]] = None,
+        check_cc: bool = True,
+        check_uc_volume: bool = True,
+        uc_volume_tolerance: float = 1e-6,
+        check_orbital_ratio: bool = True,
+    ) -> "Model":
+        """
+        Returns a model with a smaller unit cell. Orbitals which are related
+        by a lattice vector of the new unit cell are "folded" into a single
+        orbital.
+        This is the inverse operation of the supercell construction.
+
+        Parameters
+        ----------
+        new_unit_cell :
+            The unit cell of the new model. Note that this unit cell must
+            lie completely within the old unit cell.
+        unit_cell_offset :
+            Position (in cartesian coordinates) at which the new unit cell
+            is based.
+        position_tolerance :
+            Tolerance used when determining if a position mapped into the
+            new unit cell is the same as an existing orbital position.
+        orbital_labels :
+            A list of labels for the orbitals of the current model. This
+            is needed to distinguish between orbitals of the same position
+            when mapping them to the new orbitals.
+        target_indices :
+            Optional list of indices (in the current model) of the orbitals
+            which lie within the new unit cell. This can be used as a
+            check that the new orbitals are the expected ones.
+        check_cc :
+            Flag to determine whether the hoppings should be checked for
+            being perfect complex conjugates. This check can fail if the
+            model does not have exact translation symmetry w.r.t. the new
+            unit cell. If set to false, hoppings are averaged to be
+            complex conjugates.
+        check_uc_volume :
+            Flag to determine if the unit cell volume should decrease
+            by the same factor as the number of orbitals.
+        uc_volume_tolerance :
+            Absolute tolerance when checking if the unit cell volume change
+            is consistent with the change in number of orbitals.
+        check_orbital_ratio :
+            Flag to determine if the ratio of individual orbital labels
+            should be checked to be the same as the initial ratio. If
+            this is set to False, the resulting model will always
+            have ``occ=None``.
+
+        .. note :: This function is currently experimental, and its interface
+            may still change without warning.
+        """
+        if len(orbital_labels) != self.size:
+            raise ValueError(
+                f"The lenght of the 'orbital_labels' input ({len(orbital_labels)}) "
+                f"does not match the size of the model ({self.size})."
+            )
+        if self.uc is None or self.pos is None:
+            raise ValueError("Unit cell and positions must be specified for model folding.")
+
+        positions_cartesian = (self.uc.T @ self.pos.T).T
+        pos_cartesian_relative = positions_cartesian - unit_cell_offset
+        new_uc = np.array(new_unit_cell)
+        pos_reduced_new = la.solve(new_uc.T, pos_cartesian_relative.T).T
+
+        # TODO: Check that the new UC is inside the old one, or also check
+        # periodic images (w.r.t old UC) of the initial positions.
+        in_uc_indices = np.argwhere(
+            np.all(np.logical_and(pos_reduced_new >= 0, pos_reduced_new < 1), axis=-1)
+        ).flatten()
+        if target_indices is not None:
+            if not np.all(target_indices == in_uc_indices):
+                raise ValueError(
+                    f'The indices for atoms in the given unit cell ({in_uc_indices}) do not match the target indices ({target_indices}).'
+                )
+
+        # Check that all orbitals are present in the correct number
+        total_orbital_ratio = len(self.pos) / len(in_uc_indices)
+        if check_orbital_ratio:
+            orbital_counts_initial = co.Counter(orbital_labels)
+            orbital_counts_new = co.Counter(np.array(orbital_labels)[np.ix_(in_uc_indices)])
+
+            for label, count_initial in orbital_counts_initial.items():
+                count_new = orbital_counts_new.get(label, 0)
+                if not np.isclose(1 / total_orbital_ratio, count_new / count_initial):
+                    raise ValueError(
+                        "The individual orbital numbers inside the new unit cell "
+                        "are not consistent with the change in total orbital number:\n"
+                        f"Total: {len(in_uc_indices)}/{len(self.pos)}\n" + "\n".join(
+                            f"Orbital '{key}': { orbital_counts_new.get(key, 0)}/{orbital_counts_initial[key]}"
+                            for key in orbital_counts_initial.keys()
+                        )
+                    )
+        if check_uc_volume:
+            uc_volume_change_factor = la.det(self.uc) / la.det(new_uc)
+            if not np.isclose(
+                uc_volume_change_factor, total_orbital_ratio, atol=uc_volume_tolerance
+            ):
+                raise ValueError(
+                    f"The unit cell volume decreased by a factor {uc_volume_change_factor}, "
+                    "which is inconsistent with the decrease in the number of "
+                    f"orbitals (factor {total_orbital_ratio})."
+                )
+
+        new_pos = pos_reduced_new[np.ix_(in_uc_indices)]
+        in_uc_labels = np.array(orbital_labels)[np.ix_(in_uc_indices)]
+
+        new_occ: ty.Optional[int]
+        if check_orbital_ratio and self.occ is not None:
+            new_occ_float = self.occ / total_orbital_ratio
+            new_occ = int(np.round(new_occ_float))
+            if not np.isclose(new_occ, new_occ_float):
+                raise ValueError(
+                    f"The occupation number {new_occ_float} of the resulting model is fractional."
+                )
+        else:
+            new_occ = None
+
+        offset_stencil = np.array(list(itertools.product(*[range(-1, 2) for _ in range(self.dim)])))
+
+        def get_min_distance_and_offset(pos1, pos2):
+            """
+            pos1 and pos2 both reduced, within [0, 1)
+            offset in reduced, added to pos2
+            """
+
+            diff = (pos2 - pos1)
+            total_diffs = offset_stencil + diff
+            distances = la.norm(new_uc.T @ total_diffs.T, axis=0)
+            min_dist_idx = np.argmin(distances)
+            min_dist = distances[min_dist_idx]
+            min_offset = offset_stencil[np.argmin(distances)]
+            return min_dist, min_offset
+
+        def get_matching_idx_and_offset(pos_reduced, orbital_label):
+            res_idx = []
+            main_offset = np.array(np.floor(pos_reduced), dtype=int)
+            pos_relative = pos_reduced % 1
+            assert np.allclose(main_offset + pos_relative, pos_reduced)
+            for i, (pos, target_label) in enumerate(zip(new_pos, in_uc_labels)):
+                if orbital_label == target_label:
+                    dist, curr_offset = get_min_distance_and_offset(pos1=pos, pos2=pos_relative)
+
+                    if dist < position_tolerance:
+                        res_idx.append((i, main_offset - curr_offset))
+
+            if len(res_idx) == 1:
+                return res_idx[0]
+            elif len(res_idx) > 1:
+                raise ValueError(f'More than one matching index found: {res_idx}.')
+            else:
+                return None, None
+
+        new_size = len(in_uc_labels)
+        new_hop: HoppingType = co.defaultdict(lambda: np.zeros((new_size, new_size), dtype=complex))
+        for input_R, input_mat in self.hop.items():
+            R_reduced_to_new_uc = la.solve(new_uc.T, (self.uc.T @ input_R))
+            input_mat = self._array_cast(input_mat)
+            # Construct the new hoppings from the full set of current
+            # hoppings, because +R and -R do not necessarily map to the
+            # same new current_offset.
+            for effective_R, effective_input_mat in [
+                (R_reduced_to_new_uc, input_mat),
+                (-R_reduced_to_new_uc, input_mat.conjugate().transpose())
+            ]:
+                if not check_cc:
+                    effective_input_mat = effective_input_mat / 2
+                for orbital_2_idx in range(self.size):
+                    full_reduced_pos_2 = effective_R + pos_reduced_new[orbital_2_idx]
+                    new_2_idx, current_offset = get_matching_idx_and_offset(
+                        full_reduced_pos_2, orbital_labels[orbital_2_idx]
+                    )
+                    if new_2_idx is not None:
+                        new_hop[tuple(current_offset)][:, new_2_idx] += effective_input_mat[
+                            in_uc_indices, orbital_2_idx]
+
+        return Model(
+            **co.ChainMap(
+                dict(
+                    hop=new_hop,
+                    size=new_size,
+                    pos=new_pos,
+                    uc=new_uc,
+                    contains_cc=check_cc,
+                    occ=new_occ
+                ), self._input_kwargs
+            )
+        )
+
     def __add__(self, model: "Model") -> "Model":
         """
         Adds two models together by adding their hopping terms.
@@ -1250,7 +1581,9 @@ class Model(HDF5Enabled):
         # check if the occupation number matches
         if self.occ != model.occ:
             raise ValueError(
-                'Error when adding Models: occupation numbers ({0}, {1}) don\'t match'.format(self.occ, model.occ)
+                'Error when adding Models: occupation numbers ({0}, {1}) don\'t match'.format(
+                    self.occ, model.occ
+                )
             )
 
         # check if the size of the hopping matrices match
@@ -1264,9 +1597,8 @@ class Model(HDF5Enabled):
         # check if the unit cells match
         if not _check_compatibility.check_uc(self, model):
             raise ValueError(
-                'Error when adding Models: unit cells don\'t match.\nModel 1:\n{0.uc}\n\nModel 2:\n{1.uc}'.format(
-                    self, model
-                )
+                'Error when adding Models: unit cells don\'t match.\nModel 1:\n{0.uc}\n\nModel 2:\n{1.uc}'
+                .format(self, model)
             )
 
         # check if the positions match
@@ -1281,9 +1613,8 @@ class Model(HDF5Enabled):
                     break
         if not pos_match:
             raise ValueError(
-                'Error when adding Models: positions don\'t match.\nModel 1:\n{0.pos}\n\nModel 2:\n{1.pos}'.format(
-                    self, model
-                )
+                'Error when adding Models: positions don\'t match.\nModel 1:\n{0.pos}\n\nModel 2:\n{1.pos}'
+                .format(self, model)
             )
 
         # ---- MAIN PART ----