+++ tests/test_cli_parse.py
@@ -21,11 +21,14 @@ def test_cli_parse(models_equal, prefix, sample, pos_kind):
     runner = CliRunner()
     with tempfile.NamedTemporaryFile() as out_file:
         run = runner.invoke(
-            cli, ['parse', '-o', out_file.name, '-f',
-                  sample(''), '-p', prefix, '--pos-kind', pos_kind],
+            cli,
+            ['parse', '-o', out_file.name, '-f',
+             sample(''), '-p', prefix, '--pos-kind', pos_kind],
             catch_exceptions=False
         )
         print(run.output)
         model_res = tbmodels.Model.from_hdf5_file(out_file.name)
-    model_reference = tbmodels.Model.from_wannier_folder(folder=sample(''), prefix=prefix, pos_kind=pos_kind)
+    model_reference = tbmodels.Model.from_wannier_folder(
+        folder=sample(''), prefix=prefix, pos_kind=pos_kind
+    )
     models_equal(model_res, model_reference)