+++ tests/test_constructors.py
@@ -98,7 +98,12 @@ def test_from_hop_list(get_model, models_equal, sparse):
         hoppings.append([t2, 0, 0, R])
         hoppings.append([-t2, 1, 1, R])
     model1 = tbmodels.Model.from_hop_list(
-        hop_list=hoppings, contains_cc=False, on_site=(1, -1), occ=1, pos=((0., ) * 3, (0.5, 0.5, 0.)), sparse=sparse
+        hop_list=hoppings,
+        contains_cc=False,
+        on_site=(1, -1),
+        occ=1,
+        pos=((0., ) * 3, (0.5, 0.5, 0.)),
+        sparse=sparse
     )
     model2 = get_model(t1, t2, sparse=sparse)
     models_equal(model1, model2)
@@ -125,7 +130,12 @@ def test_from_hop_list_with_cc(get_model, models_close, sparse):
         hoppings.append([-t2, 1, 1, R])
         hoppings.append([-t2, 1, 1, tuple(-x for x in R)])
     model1 = tbmodels.Model.from_hop_list(
-        hop_list=hoppings, contains_cc=True, on_site=(1, -1), occ=1, pos=((0., ) * 3, (0.5, 0.5, 0.)), sparse=sparse
+        hop_list=hoppings,
+        contains_cc=True,
+        on_site=(1, -1),
+        occ=1,
+        pos=((0., ) * 3, (0.5, 0.5, 0.)),
+        sparse=sparse
     )
     model2 = get_model(t1, t2, sparse=sparse)
     models_close(model1, model2)
@@ -146,7 +156,12 @@ def test_pos_outside_uc_with_hoppings(get_model, models_equal, sparse):  # pylin
         hoppings.append([t2, 0, 0, R])
         hoppings.append([-t2, 1, 1, R])
     model1 = tbmodels.Model.from_hop_list(
-        hop_list=hoppings, contains_cc=False, on_site=(1, -1), occ=1, pos=((0., ) * 3, (-0.5, -0.5, 0.)), sparse=sparse
+        hop_list=hoppings,
+        contains_cc=False,
+        on_site=(1, -1),
+        occ=1,
+        pos=((0., ) * 3, (-0.5, -0.5, 0.)),
+        sparse=sparse
     )
     model2 = get_model(t1, t2, sparse=sparse)
     models_equal(model1, model2)
@@ -176,7 +191,11 @@ def test_non_hermitian_2():
     to a hermitian Hamiltonian.
     """
     with pytest.raises(ValueError):
-        tbmodels.Model(size=2, hop={(0, 0, 0): np.eye(2), (1, 0, 0): np.eye(2), (-1, 0, 0): 2 * np.eye(2)})
+        tbmodels.Model(
+            size=2, hop={(0, 0, 0): np.eye(2),
+                         (1, 0, 0): np.eye(2),
+                         (-1, 0, 0): 2 * np.eye(2)}
+        )
 
 
 def test_wrong_key_length():
@@ -186,9 +205,11 @@ def test_wrong_key_length():
     """
     with pytest.raises(ValueError):
         tbmodels.Model(
-            size=2, hop={(0, 0, 0): np.eye(2),
-                         (1, 0, 0): np.eye(2),
-                         (-1, 0, 0, 0): np.eye(2)}, contains_cc=False
+            size=2,
+            hop={(0, 0, 0): np.eye(2),
+                 (1, 0, 0): np.eye(2),
+                 (-1, 0, 0, 0): np.eye(2)},
+            contains_cc=False
         )